app/requirements.txt
amqp==5.0.9
autopep8==1.6.0
billiard==3.6.4.0
blinker==1.4
cachetools==5.0.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect==1.1.0
cssutils==2.3.0
Deprecated==1.2.13
ffmpeg-python==0.2.0
Flask==2.0.2
flask-buzz==2.0.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
flask-praetorian==1.3.0
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.2
idna==3.3
inflection==0.5.1
itsdangerous==2.0.1
Jinja2==3.0.3
kombu==5.2.3
lxml==4.7.1
MarkupSafe==2.0.1
marshmallow==3.14.1
packaging==21.3
passlib==1.7.4
pendulum==2.1.2
premailer==3.10.0
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
py-buzz==2.1.3
pycodestyle==2.8.0
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytzdata==2020.1
redis==4.1.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
toml==0.10.2
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.13.3
yagmail==0.15.277
